Use google CSE to search docs
google CSE (Custom Search) is similar to searching google
for a particular website
eg. search query "site:doc.crates.io custom"

It is a service hosted by Google. It has a free plan
which serves ads. And paid plans without ads and
more customization evidently (have not looked into it)

but we can

- include multiple websites
- has various options for styling

But the design of the results page looks vintage
and not in a good way

the advantage is that it is free of course

there are multiple options for integrating google CSE
here we

- have our own search box. which we have styled similar
  to search crates
- searching opens the results in a new tab

normally google CSE, asks you to add their javascript
but we make do with redirecting to their url

for more information

https://support.google.com/customsearch/answer/4513751?hl=en
https://en.wikipedia.org/wiki/Google_Custom_Search
https://cse.google.com